+++ tests/test_wavtrans.py
@@ -361,7 +361,7 @@ class TestWavTrans(unittest.TestCase):
 
         alm = np.asarray([1, 2, 3, 4, 5, 6, 7, 8, 9, 10], dtype=np.complex64)
         ainfo = sharp.alm_info(lmax=3)
-        w_ell = np.zeros((2, 4))
+        w_ell = np.zeros((2, 4), dtype=np.float32)
         w_ell[0,:2] = 1
         w_ell[1,2:] = 0.5
         spin = 0
@@ -424,7 +424,7 @@ class TestWavTrans(unittest.TestCase):
 
         alm = np.asarray([1, 2, 3, 4, 5, 6, 7, 8, 9, 10], dtype=np.complex64)
         ainfo = sharp.alm_info(lmax=3)
-        w_ell = np.zeros((2, 4))
+        w_ell = np.zeros((2, 4), dtype=np.float32)
         w_ell[0,:2] = 1
         w_ell[1,2:] = 0.5
         spin = 0
@@ -453,7 +453,7 @@ class TestWavTrans(unittest.TestCase):
 
         alm = np.asarray([1, 2, 3, 4, 5, 6, 7, 8, 9, 10], dtype=np.complex64)
         ainfo = sharp.alm_info(lmax=3)
-        w_ell = np.zeros((2, 4))
+        w_ell = np.zeros((2, 4), dtype=np.float32)
         w_ell[0,:2] = 1
         w_ell[1,2:] = 0.5
         spin = 0
@@ -481,7 +481,7 @@ class TestWavTrans(unittest.TestCase):
 
         alm = np.asarray([1, 2, 3, 4, 5, 6, 7, 8, 9, 10], dtype=np.complex64)
         ainfo = sharp.alm_info(lmax=3)
-        w_ell = np.zeros((2, 4))
+        w_ell = np.zeros((2, 4), dtype=np.float32)
         w_ell[0,:2] = 1
         w_ell[1,2:] = 0.5
         spin = 0
